backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
openai==1.54.3
python-dotenv==1.0.1
PySocks==1.7.1
websockets==13.1
boto3==1.35.0
python-telegram-bot==21.0
requests==2.31.0
motor<3.6
pymongo<4.9
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
aiokafka==0.10.0
confluent-kafka==2.3.0
httpx==0.27.0
email-validator

# Redis для кэширования
redis==5.0.1
aioredis==2.0.1

# Celery для фоновых задач
celery==5.3.4
flower==2.0.1

# Мониторинг и метрики
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# Circuit Breaker
pybreaker==1.0.1

# Rate Limiting
slowapi==0.1.9

# Структурированное логирование
structlog==24.1.0

# Улучшенная обработка ошибок
tenacity==8.2.3

# Библиотеки для OCR
pytesseract==0.3.10
pillow==10.2.0